requirements.txt
moderngl
PyGLM
PyGame
PyOpenGL
numpy